fix(category): change all entity and domain id types from Long to Integer to match INT columns

+++ TuEventoBackend/tu-evento/src/main/java/com/capysoft/tuevento/modules/category/application/dto/request/AssignCategoryRequest.java
@@ -11,8 +11,8 @@ import lombok.Setter;
 public class AssignCategoryRequest {
 
     @NotNull
-    private Long categoryId;
+    private Integer categoryId;
 
     @NotNull
-    private Long eventId;
+    private Integer eventId;
 }
