+++ TuEventoBackend/tu-evento/src/main/java/com/capysoft/tuevento/modules/category/application/dto/response/CategoryResponse.java
@@ -7,10 +7,10 @@ import lombok.Getter;
 @Builder
 public class CategoryResponse {
 
-    private Long    categoryId;
+    private Integer categoryId;
     private String  name;
     private String  description;
     private boolean active;
     private boolean visible;
-    private Long    dadId;
+    private Integer dadId;
 }
