+++ TuEventoBackend/tu-evento/src/main/java/com/capysoft/tuevento/modules/category/domain/repository/CategoryRepository.java
@@ -9,21 +9,21 @@ public interface CategoryRepository {
 
     Category save(Category category);
 
-    Optional<Category> findById(Long id);
+    Optional<Category> findById(Integer id);
 
     List<Category> findAll();
 
     List<Category> findAllActive();
 
     /** Returns all subcategories whose parent is {@code dadId}. */
-    List<Category> findByDadId(Long dadId);
+    List<Category> findByDadId(Integer dadId);
 
     /** Returns all root categories (dadId IS NULL). */
     List<Category> findRootCategories();
 
     boolean existsByName(String name);
 
-    boolean existsById(Long id);
+    boolean existsById(Integer id);
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 }
